Add 'play' method to allow for playing of sounds.
There is no ability to send sounds to Campfire directly from the
response message object. However, using Campfire you simply "/play
sound".

Add this ability to the `Adapter` and the provided default adapter
implementations. It is added to the `Shell` adapter to allow for an
easier time at testing / debugging scripts using the local shell.

Add to the `Campfire` adapter as this is the core place where it
can be used.