Determine adapterPath in robot.coffee, rather than bin/hubot
Previously, adapterPath was determined in bin/hubot, and passed to index.coffee,
 and eventually passed ot the constructor in robot.coffee. This is used by Robot
to determine where to load builtin adapters from.

It also makes it hard to progmatically construct a robot without running bin/hubot,
since you'd need to find a way to figure out the path that the hubot module is installed.

This changes it so Robot will determine the adapterPath since it's relative to where
it live. The constructor still takes it as an option so it remains backward compatible
 to anything that might be specifying it.