docs(getting started): clarify hubot name in commands
Following the current getting started guide, the name of the hubot becomes `myhubot`. Following the instructions I got stuck at the prompt `hubot> hubot help` which responded with nothing. Once I understood the problem it was clear to me, but for a first timer this might be confusing. Ideally the prompt would return with a helpful notice to `hubot help`, but for the time being, I drafted an update to the getting started docs with the goal to make things more clear